feat(BaseService.js): add static method wpaction to handle WordPress actions fix(Index.js): remove unused imports and variables, refactor delete operation function to show confirmation dialog before deleting

+++ src/views/Operations/Index.js
@@ -1,18 +1,12 @@
-import { Stack, Divider, List, Select, Button, Card, CardContent, CardHeader, Box, TextField, ListItem, ListItemText, Typography, MenuItem, IconButton, ButtonBase } from '@mui/material'
+import { Stack, Divider, List, Card, CardContent, CardHeader, ListItem, ListItemText, Typography, IconButton } from '@mui/material'
 import React, { useEffect, useState } from 'react'
 import OperationService from '../../services/OperationService'
-import CurrencyService from '../../services/CurrencyService'
 import { LoadingButton } from '@mui/lab'
-import { Add, Check, Remove, Delete } from '@mui/icons-material'
+import { Check, Delete } from '@mui/icons-material'
 
-import { handlerChange } from '../../tools/Common'
 
 export default function Index() {
-    const [ from, setFrom ] = useState('')
-    const [ to, setTo ] = useState('')
-    const [ price, setPrice ] = useState(1)
     const [ operations, setOperations ] = useState([])
-    const [ currencies, setCurrencies ] = useState([])
     const [ loading, setLoading ] = useState(false)
     const [ loaded, setLoaded ] = useState(false)
 
@@ -22,43 +16,26 @@ export default function Index() {
             (async ()=>{
                 setLoading(true)
                 const tempOperations = await OperationService.get()
-                const tempCurrencies = await CurrencyService.get()
-                console.log( tempOperations )
                 setOperations( prev => tempOperations )
-                setCurrencies( prev => tempCurrencies )
                 setLoaded(true)
                 setLoading(false)
             })()
         }
     },[loaded, operations])
 
-    const saveChange = async () => {
-        setLoading(true)
-        const tempOperations = await OperationService.set({
-            currency_from: from,
-            currency_to: to,
-            price: price,
-            rules: []
-        })
-        setOperations( prev => tempOperations )
-        setFrom('')
-        setTo('')
-        setPrice(1)
-        setLoading(false)
-    }
-
     const handlerDelete = (id) => async () => {
-        setLoading(true)
-        const result = await OperationService.delete({
-            id
-        })
-        setOperations( prev => result )
-        setLoading(false)
+        if( window.confirm("Seguro de borrar?") ) {
+            setLoading(true)
+            const result = await OperationService.delete({
+                id
+            })
+            setOperations( prev => result )
+            setLoading(false)
+        }
     }
 
     const clearAll = async () => {
-        // const response = confirm("Borrar?")
-        if( true ) {
+        if( window.confirm("Seguro de borrar?") ) {
             setLoading(true)
             for(const operation of operations ) {
                 await OperationService.delete({
@@ -82,11 +59,11 @@ export default function Index() {
     return <Card>
         <CardHeader title="Operaciones" />
         <CardContent>
-            <Button variant='outlined' onClick={clearAll}>Borrar</Button>
+            <LoadingButton variant='outlined' loading={loading} onClick={clearAll}>Borrar</LoadingButton>
             <List>{operations.map( operation => <span key={operation.id}>
                 <ListItem secondaryAction={<>
                         <IconButton onClick={()=>handlerVerify(operation.id)} disabled={operation.verify}><Check sx={{ color: operation.verify ? 'green' : 'gray' }}/></IconButton>
-                        <IconButton onClick={handlerDelete}><Delete sx={{ color: 'red' }}/></IconButton>
+                        <IconButton onClick={handlerDelete(operation.id)}><Delete sx={{ color: 'red' }}/></IconButton>
                     </>}>
                     <ListItemText 
                         primary={<Stack spacing={1}>
